requirements.txt
# 核心數據處理和分析
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.20

# 技術指標計算
TA-Lib>=0.4.26

# 測試套件
pytest>=7.0.0
pytest-cov>=4.0.0

# 程式碼品質
flake8>=6.0.0
black>=23.0.0

# 圖表和可視化
matplotlib>=3.7.0
seaborn>=0.12.0

# 資料庫連接
pyodbc
python-dotenv